test(transform): split exported aggregate regression
