+++ crates/perry-transform/src/aggregate_scalar.rs
@@ -1726,7 +1726,7 @@ mod tests {
         }
     }
 
-    fn aggregate_fixture(observe_identity: bool) -> Module {
+    pub(super) fn aggregate_fixture(observe_identity: bool) -> Module {
         let mut module = Module::new("aggregate-scalar.ts");
         module.init = vec![
             Stmt::Let {
@@ -1802,44 +1802,6 @@ mod tests {
             .init
             .iter()
             .any(|stmt| matches!(stmt, Stmt::Let { id: 2, .. })));
-    }
-
-    #[test]
-    fn exported_aggregate_keeps_materialized_carrier() {
-        for metadata_source in ["exports", "exported_objects"] {
-            let mut module = aggregate_fixture(false);
-            if metadata_source == "exports" {
-                module.exports.push(Export::Named {
-                    local: "values".to_string(),
-                    exported: "VALUES".to_string(),
-                });
-            } else {
-                module.exported_objects.push("values".to_string());
-            }
-
-            run(&mut module);
-
-            assert!(
-                module.init.iter().any(|stmt| {
-                    matches!(
-                        stmt,
-                        Stmt::Let {
-                            id: 1,
-                            init: Some(Expr::Array(_)),
-                            ..
-                        }
-                    )
-                }),
-                "{metadata_source} must keep the exported carrier"
-            );
-            assert!(
-                module
-                    .init
-                    .iter()
-                    .any(|stmt| matches!(stmt, Stmt::Let { id: 2, .. })),
-                "{metadata_source} must keep element aliases"
-            );
-        }
     }
 
     #[test]
@@ -2026,3 +1988,7 @@ mod tests {
         assert!(format!("{stmts:?}").contains("__AnonShape_short"));
     }
 }
+
+#[cfg(test)]
+#[path = "aggregate_scalar_export_tests.rs"]
+mod export_tests;
